Removing _ret for pipe relationships (#327)
A new bidirectional dictionary for the pipe relationships has been setup.
Backwards compatibility is secured by checking the ESDLversion, if older than 21.10, the former _ret pipe relation is used.
Fix on the definition for hot_pipes, this now only contains pipes that have a relation to a cold_pipe, else it is added to the list of unrelated pipes.
Be aware, unrelated pipes significantly increase the computational effort.